refactor: rename platform packages and consolidate dsn logic
