+++ internal/platform/dsn/dsn_test.go
@@ -1,4 +1,4 @@
-package common
+package dsn
 
 import "testing"
