+++ internal/platform/logging/logging_test.go
@@ -1,4 +1,4 @@
-package logger
+package logging
 
 import (
 	"strings"
